# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Sales Assistant API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: leads
    description: Lead management operations
  - name: emails
    description: Email drafting and sending operations
  - name: dashboard
    description: Dashboard and analytics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"

  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /leads/{id}/emails:
    get:
      operationId: listLeadEmails
      tags: [emails]
      summary: List emails for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of emails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Email"

  /emails:
    get:
      operationId: listEmails
      tags: [emails]
      summary: List all emails
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of emails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Email"

  /emails/draft:
    post:
      operationId: draftEmail
      tags: [emails]
      summary: AI-draft a personalized outreach email for a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftEmailBody"
      responses:
        "201":
          description: Created email draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"

  /emails/{id}:
    get:
      operationId: getEmail
      tags: [emails]
      summary: Get an email by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateEmail
      tags: [emails]
      summary: Update an email (subject/body edits before sending)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmailBody"
      responses:
        "200":
          description: Updated email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteEmail
      tags: [emails]
      summary: Delete an email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /emails/{id}/send:
    post:
      operationId: sendEmail
      tags: [emails]
      summary: Mark email as sent (simulate sending)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email marked as sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard summary statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      responses:
        "200":
          description: Recent activity items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /dashboard/lead-status-breakdown:
    get:
      operationId: getLeadStatusBreakdown
      tags: [dashboard]
      summary: Get lead counts grouped by status
      responses:
        "200":
          description: Status breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusCount"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Lead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        company:
          type: string
        jobTitle:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - company
        - status
        - createdAt
        - updatedAt

    CreateLeadBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        company:
          type: string
        jobTitle:
          type: string
        industry:
          type: string
        website:
          type: string
        notes:
          type: string
        status:
          type: string
      required:
        - name
        - email
        - company

    UpdateLeadBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        company:
          type: string
        jobTitle:
          type: string
        industry:
          type: string
        website:
          type: string
        notes:
          type: string
        status:
          type: string

    Email:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
        subject:
          type: string
        body:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        sentAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - leadId
        - subject
        - body
        - status
        - createdAt

    DraftEmailBody:
      type: object
      properties:
        leadId:
          type: integer
        context:
          type: string
          description: Optional extra context or sales angle to guide the AI
      required:
        - leadId

    UpdateEmailBody:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string

    DashboardStats:
      type: object
      properties:
        totalLeads:
          type: integer
        emailsSent:
          type: integer
        emailsDrafted:
          type: integer
        openRate:
          type: number
        newLeadsThisWeek:
          type: integer
      required:
        - totalLeads
        - emailsSent
        - emailsDrafted
        - openRate
        - newLeadsThisWeek

    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        leadName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - description
        - leadName
        - createdAt

    StatusCount:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
      required:
        - status
        - count

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error