# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Sales Assistant API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all leads
 */
export const ListLeadsQueryParams = zod.object({
  status: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
});

export const ListLeadsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  company: zod.string(),
  jobTitle: zod.string().nullish(),
  industry: zod.string().nullish(),
  website: zod.string().nullish(),
  notes: zod.string().nullish(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListLeadsResponse = zod.array(ListLeadsResponseItem);

/**
 * @summary Create a new lead
 */
export const CreateLeadBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  company: zod.string(),
  jobTitle: zod.string().optional(),
  industry: zod.string().optional(),
  website: zod.string().optional(),
  notes: zod.string().optional(),
  status: zod.string().optional(),
});

/**
 * @summary Get a lead by ID
 */
export const GetLeadParams = zod.object({
  id: zod.coerce.number(),
});

export const GetLeadResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  company: zod.string(),
  jobTitle: zod.string().nullish(),
  industry: zod.string().nullish(),
  website: zod.string().nullish(),
  notes: zod.string().nullish(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a lead
 */
export const UpdateLeadParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateLeadBody = zod.object({
  name: zod.string().optional(),
  email: zod.string().optional(),
  company: zod.string().optional(),
  jobTitle: zod.string().optional(),
  industry: zod.string().optional(),
  website: zod.string().optional(),
  notes: zod.string().optional(),
  status: zod.string().optional(),
});

export const UpdateLeadResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  company: zod.string(),
  jobTitle: zod.string().nullish(),
  industry: zod.string().nullish(),
  website: zod.string().nullish(),
  notes: zod.string().nullish(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a lead
 */
export const DeleteLeadParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List emails for a lead
 */
export const ListLeadEmailsParams = zod.object({
  id: zod.coerce.number(),
});

export const ListLeadEmailsResponseItem = zod.object({
  id: zod.number(),
  leadId: zod.number(),
  subject: zod.string(),
  body: zod.string(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
  sentAt: zod.coerce.date().nullish(),
});
export const ListLeadEmailsResponse = zod.array(ListLeadEmailsResponseItem);

/**
 * @summary List all emails
 */
export const ListEmailsQueryParams = zod.object({
  status: zod.coerce.string().optional(),
});

export const ListEmailsResponseItem = zod.object({
  id: zod.number(),
  leadId: zod.number(),
  subject: zod.string(),
  body: zod.string(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
  sentAt: zod.coerce.date().nullish(),
});
export const ListEmailsResponse = zod.array(ListEmailsResponseItem);

/**
 * @summary AI-draft a personalized outreach email for a lead
 */
export const DraftEmailBody = zod.object({
  leadId: zod.number(),
  context: zod
    .string()
    .optional()
    .describe("Optional extra context or sales angle to guide the AI"),
});

/**
 * @summary Get an email by ID
 */
export const GetEmailParams = zod.object({
  id: zod.coerce.number(),
});

export const GetEmailResponse = zod.object({
  id: zod.number(),
  leadId: zod.number(),
  subject: zod.string(),
  body: zod.string(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
  sentAt: zod.coerce.date().nullish(),
});

/**
 * @summary Update an email (subject/body edits before sending)
 */
export const UpdateEmailParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateEmailBody = zod.object({
  subject: zod.string().optional(),
  body: zod.string().optional(),
});

export const UpdateEmailResponse = zod.object({
  id: zod.number(),
  leadId: zod.number(),
  subject: zod.string(),
  body: zod.string(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
  sentAt: zod.coerce.date().nullish(),
});

/**
 * @summary Delete an email
 */
export const DeleteEmailParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Mark email as sent (simulate sending)
 */
export const SendEmailParams = zod.object({
  id: zod.coerce.number(),
});

export const SendEmailResponse = zod.object({
  id: zod.number(),
  leadId: zod.number(),
  subject: zod.string(),
  body: zod.string(),
  status: zod.string(),
  createdAt: zod.coerce.date(),
  sentAt: zod.coerce.date().nullish(),
});

/**
 * @summary Get dashboard summary statistics
 */
export const GetDashboardStatsResponse = zod.object({
  totalLeads: zod.number(),
  emailsSent: zod.number(),
  emailsDrafted: zod.number(),
  openRate: zod.number(),
  newLeadsThisWeek: zod.number(),
});

/**
 * @summary Get recent activity feed
 */
export const GetRecentActivityResponseItem = zod.object({
  id: zod.number(),
  type: zod.string(),
  description: zod.string(),
  leadName: zod.string(),
  createdAt: zod.coerce.date(),
});
export const GetRecentActivityResponse = zod.array(
  GetRecentActivityResponseItem,
);

/**
 * @summary Get lead counts grouped by status
 */
export const GetLeadStatusBreakdownResponseItem = zod.object({
  status: zod.string(),
  count: zod.number(),
});
export const GetLeadStatusBreakdownResponse = zod.array(
  GetLeadStatusBreakdownResponseItem,
);